Remove unused federation endpoint (`query_auth`) (#7026)
