+++ changelog.d/7026.removal
@@ -1,0 +1,1 @@
+Remove the unused query_auth federation endpoint per MSC2451.
